Edit view livewire crm customer-group-lists

+++ resources/views/livewire/customer-group/customer-group-lists.blade.php
@@ -39,15 +39,17 @@
                 </div>
                 <div class="col-3 d-flex justify-content-center">
                     <div class="btn-group w-100" role="group">
-                        <button wire:click="$dispatch('add-customer-group')" type="button" class="btn btn-primary"
-                            data-toggle="modal" data-target="#modal-add-customer-group">
-                            <i class="fas fa-plus"></i> Customer Group
-                        </button>
+
+                        @can('customerGroup.create')
+                            <button wire:click="$dispatch('add-customer-group')" type="button" class="btn btn-primary"
+                                data-toggle="modal" data-target="#modal-add-customer-group">
+                                <i class="fas fa-plus"></i> Customer Group
+                            </button>
+                        @endcan
+
                         <button wire:click="$dispatch('refresh-customer-group')" type="button" class="btn btn-success">
                             <i class="fas fa-sync-alt"></i> Refresh
                         </button>
-
-                        @livewire('customer-group.customer-group-create')
                     </div>
                 </div>
             </div>
@@ -68,14 +70,18 @@
         <!-- /.card-header -->
         <div class="card-body">
             <div class="table-responsive">
-                <table class="table table-hover">
-                    <thead>
+                <table class="table table-sm table-hover table-bordered">
+                    <thead class="table-info">
                         <th scope="col">#</th>
                         <th scope="col">Created</th>
                         <th scope="col">Updated</th>
                         <th scope="col">Id</th>
                         <th scope="col">Customer Group Name</th>
-                        <th scope="col" style="width: 100px">Action</th>
+                        @can('customerGroup.edit')
+                            <th scope="col" colspan="2">Action</th>
+                        @elsecan('customerGroup.delete')
+                            <th scope="col" colspan="2">Action</th>
+                        @endcan
                     </thead>
 
                     <tbody>
@@ -120,24 +126,36 @@
                                 </td>
                                 <td>{{ $item->id }}</td>
                                 <td>{{ $item->name }}</td>
-                                <td>
-                                    <button
-                                        wire:click.prevent="$dispatch('edit-customer-group',{id:{{ $item->id }}})"
-                                        type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
-                                        data-target="#modal-edit-customer-group">
-                                        <i class="fas fa-edit"></i>
-                                    </button>
-                                    <button
-                                        wire:click.prevent="deleteCustomerGroup({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
-                                        class="btn btn-sm btn-danger">
-                                        <i class="fas fa-trash"></i>
-                                    </button>
-                                </td>
+
+                                @can('customerGroup.edit')
+                                    <td style="width: 40px" class="p-1 text-center">
+                                        <button
+                                            wire:click.prevent="$dispatch('edit-customer-group',{id:{{ $item->id }}})"
+                                            type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
+                                            data-target="#modal-edit-customer-group">
+                                            <i class="fas fa-edit"></i>
+                                        </button>
+                                    </td>
+                                @endcan
+
+                                @can('customerGroup.delete')
+                                    <td style="width: 40px" class="p-1 text-center">
+                                        <button
+                                            wire:click.prevent="deleteCustomerGroup({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
+                                            class="btn btn-sm btn-danger">
+                                            <i class="fas fa-trash"></i>
+                                        </button>
+                                    </td>
+                                @endcan
                             </tr>
                         @endforeach
-                        @livewire('customer-group.customer-group-edit')
                     </tbody>
                 </table>
+
+                @livewire('customer-group.customer-group-create')
+
+                @livewire('customer-group.customer-group-edit')
+
             </div>
         </div>
         <!-- /.card-body -->
